[FEAT(Player)] Dealer와 User의 공통메소드 추상화
